requirements.txt
pyyaml
flask
requests
prometheus_client
uwsgi

werkzeug==0.16.0
py-healthcheck
